Change Docker config / setup and composer wait for accepting mysql connect
