+++ Makefile
@@ -7,7 +7,10 @@ IMAGE := $(DOCKER_REGISTRY)/$(REPOSITORY)/$(NAME)
 
 test:
 	true
-
+build-binary:
+	cd src && go mod download && CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo  -o ../bin/go-boilerplate
+run-docker: build-binary
+	docker-compose up --build
 image:
 	docker build -t $(IMAGE):$(VERSION) .
