--- init/01.sql
@@ -1,9 +1,0 @@
-CREATE TABLE students (
-    id int,
-    name VARCHAR(255)
-);
-
-INSERT INTO students(id, name) VALUES
-(1,'A'),
-(2,'B'),
-(3,'C');
