--- model/student.go
@@ -1,6 +1,0 @@
-package model
-
-type Students struct {
-	Id   string `json:"id"`
-	Name string `json:"name"`
-}
