+++ src/model/student.go
@@ -1,0 +1,6 @@
+package model
+
+type Students struct {
+	Id   string `json:"id"`
+	Name string `json:"name"`
+}
